fix: fix code errors after failing tests

+++ python_variable_annotations/7-to_kv.py
@@ -7,4 +7,4 @@ def to_kv(k: str, v: Union[int, float]) -> Tuple[str, float]:
     """takes a string k and an int OR float v as arguments
         and returns a tuple.
     """
-    return (k, float(v * v))
+    return (k, v * v)
